users/requirements.txt
Django==5.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.2.1
Celery==5.6.2
